+++ Backend-Task-page/utils/auth.py
@@ -1,5 +1,5 @@
-from fastapi import Request, HTTPException
-from jose import jwt, JWTError
+from fastapi import HTTPException, Request
+from jose import JWTError, jwt
 
 SECRET_KEY = "mytask73391638281111042110@!@#$%^&*()_+=-<>?/.,][{]"
 ALGORITHM = "HS256"
@@ -16,9 +16,26 @@ def get_current_user(request: Request):
         role = payload.get("role")
         if username is None:
             raise HTTPException(status_code=401, detail="Invalid token payload")
-        return {
-            "username": username,
-            "role" : role
-            }
+        return {"username": username, "role": role}
+    except JWTError:
+        raise HTTPException(status_code=401, detail="Invalid token")
+
+
+def get_current_admin(request: Request):
+    auth_header = request.headers.get("Authorization")
+    if not auth_header or not auth_header.startswith("Bearer "):
+        raise HTTPException(status_code=401, detail="Invalid or missing token")
+
+    token = auth_header.split(" ")[1]
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        username = payload.get("sub")
+        if username is None:
+            raise HTTPException(status_code=401, detail="Invalid token payload")
+        from database import user_collection
+        found_admin = user_collection.find_one({"username": username})
+        if not found_admin or found_admin["role"] != "admin":
+            raise HTTPException(status_code=401, detail="Unauthorized Admin")
+        return {"username": username}
     except JWTError:
         raise HTTPException(status_code=401, detail="Invalid token")
