+++ Frontend-Task-page/src/Components/AdminSidebar.jsx
@@ -10,6 +10,35 @@ export default function AdminSidebar({onLogout}) {
     const [isMobile, setIsMobile] = useState(false);
     const navigate = useNavigate();
     const location = useLocation();
+    const role = localStorage.getItem("role")
+    const rolePaths = {
+        home : {
+            admin : "/admin/Home",
+            HR : "/hr/Home",
+            Manager : "/manager/Home",
+            Designer : "/user/Home",
+            Worker : "/user/Home"
+        },
+        clientView : {
+            admin : "/admin/clientView",
+            HR : "/hr/clientView",
+            Manager : "/manager/clientView",
+            Designer : "/user/clientView"
+        },
+        createStaff : {
+            admin : "/admin/createStaff",
+            HR : "/hr/createStaff",
+        },
+        viewStaffDetails : {
+            admin : "/admin/viewStaffDetails",
+            HR : "/hr/viewStaffDetails",
+            Manager : "/manager/viewStaffDetails"
+        },
+        attendance : {
+            Designer : "/user/attendance",
+            Worker : "/user/attendance"
+        }
+    }
 
     useEffect(() => {
         const handleResize = () => {
@@ -28,7 +57,7 @@ export default function AdminSidebar({onLogout}) {
     return(
         <nav className={styles.sidebar}>
             <ul className={styles.Container}>
-                <li className={`${styles.home}`} onClick={()=> navigate('/admin/adminhome')}>
+                <li className={`${styles.home}`} onClick={() => navigate(rolePaths.home[role])}>
                     <a className={styles.h_icon}>
                         <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 37 33" fill="none">
                             <path d="M5.09375 13.1484V31.125C5.09375 31.4482 5.22215 31.7582 5.45071 31.9868C5.67927 32.2153 5.98927 32.3438 6.3125 32.3438H13.625V21.9844C13.625 21.4995 13.8176 21.0345 14.1604 20.6917C14.5033 20.3489 14.9683 20.1562 15.4531 20.1562H21.5469C22.0317 20.1562 22.4967 20.3489 22.8396 20.6917C23.1824 21.0345 23.375 21.4995 23.375 21.9844V32.3438H30.6875C31.0107 32.3438 31.3207 32.2153 31.5493 31.9868C31.7778 31.7582 31.9062 31.4482 31.9062 31.125V13.1484" stroke="white" stroke-linecap="round" stroke-linejoin="round"/>
@@ -36,28 +65,67 @@ export default function AdminSidebar({onLogout}) {
                         </svg>
                     </a>
                 </li>
-                <li className={`${styles.entry}`} onClick={() => navigate("/admin/clientview")}>
+                <li className={`${styles.entry}`} onClick={() => navigate(rolePaths.clientView[role])}>
                     <a className={styles.e_icon}>
-                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32" fill="none">
-                            <path d="M22.4004 0.5C22.4798 0.500103 22.5561 0.531722 22.6123 0.587891C22.6685 0.644096 22.7001 0.720355 22.7002 0.799805V3.7002H28C28.9283 3.7002 29.8182 4.06921 30.4746 4.72559C31.1309 5.38195 31.5 6.27198 31.5 7.2002V28C31.5 28.9283 31.131 29.8182 30.4746 30.4746C29.8182 31.131 28.9283 31.5 28 31.5H4.00098C3.0729 31.4989 2.18262 31.1299 1.52637 30.4736C0.911122 29.8584 0.548749 29.0377 0.504883 28.1729L0.5 27.999V7.2002L0.504883 7.02734C0.548702 6.16244 0.911136 5.34184 1.52637 4.72656C2.14162 4.11131 2.96225 3.74796 3.82715 3.7041L4.00098 3.7002H9.2998V0.799805C9.29986 0.720357 9.33154 0.644096 9.3877 0.587891C9.44386 0.531722 9.52019 0.500103 9.59961 0.5C9.67917 0.5 9.75624 0.53163 9.8125 0.587891C9.86859 0.644086 9.90034 0.720401 9.90039 0.799805V3.7002H22.0996V0.799805C22.0997 0.720401 22.1314 0.644086 22.1875 0.587891C22.2438 0.53163 22.3208 0.5 22.4004 0.5ZM1.09961 28C1.09961 28.7691 1.40536 29.5069 1.94922 30.0508C2.49307 30.5946 3.23087 30.9004 4 30.9004H28C28.7691 30.9004 29.5069 30.5946 30.0508 30.0508C30.5946 29.5069 30.9004 28.7691 30.9004 28V13.9004H1.09961V28ZM4 4.2998C2.39542 4.2998 1.09961 5.59969 1.09961 7.2002V13.2998H30.9004V7.2002C30.9004 6.43111 30.5946 5.69326 30.0508 5.14941C29.5069 4.60556 28.7691 4.2998 28 4.2998H22.7002V7.2002C22.7001 7.27964 22.6685 7.3559 22.6123 7.41211C22.5561 7.46828 22.4798 7.4999 22.4004 7.5C22.3208 7.5 22.2438 7.46837 22.1875 7.41211C22.1314 7.35591 22.0997 7.2796 22.0996 7.2002V4.2998H9.90039V7.2002C9.90034 7.2796 9.86859 7.35591 9.8125 7.41211C9.75624 7.46837 9.67917 7.5 9.59961 7.5C9.52019 7.4999 9.44386 7.46828 9.3877 7.41211C9.33154 7.3559 9.29986 7.27964 9.2998 7.2002V4.2998H4Z" fill="white" stroke="white"/>
-                            <path d="M15.3924 17.2001H15.4084V17.2171H15.3924V17.2001ZM15.3924 21.9916H15.4084V22.0086H15.3924V21.9916ZM15.3924 26.783H15.4084V26.8001H15.3924V26.783ZM10.8964 21.9916H10.9124V22.0086H10.8964V21.9916ZM10.8964 26.783H10.9124V26.8001H10.8964V26.783ZM6.40039 21.9916H6.41638V22.0086H6.40039V21.9916ZM6.40039 26.783H6.41638V26.8001H6.40039V26.783ZM19.8884 17.2001H19.9044V17.2171H19.8884V17.2001ZM19.8884 21.9916H19.9044V22.0086H19.8884V21.9916ZM19.8884 26.783H19.9044V26.8001H19.8884V26.783ZM24.3844 17.2001H24.4004V17.2171H24.3844V17.2001ZM24.3844 21.9916H24.4004V22.0086H24.3844V21.9916Z" fill="white" stroke="white" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/>
+                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="20" viewBox="0 0 16 20" fill="none">
+                            <path d="M8 11V15M6 13H10M3 19C2.46957 19 1.96086 18.7893 1.58579 18.4142C1.21071 18.0391 1 17.5304 1 17V1H10L15 6V17C15 17.5304 14.7893 18.0391 14.4142 18.4142C14.0391 18.7893 13.5304 19 13 19H3Z" stroke="white" stroke-linecap="round" stroke-linejoin="round"/>
+                            <path d="M9 1V7H15" stroke="white" stroke-linejoin="round"/>
                         </svg>
                     </a>
                 </li>
-                <li className={`${styles.entry}`} onClick={() => navigate("/admin/signupstaff")}>
-                    <a className={styles.e_icon}>
-                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 34 34" fill="none">
-                            <path d="M15.1469 0C20.2414 0 24.3711 4.10562 24.3711 9.17125C24.3723 10.4987 24.0836 11.8104 23.5254 13.0147C22.9672 14.2191 22.1527 15.2871 21.139 16.1441C21.571 16.3574 21.9646 16.5724 22.3198 16.7889C23.0836 17.2529 23.8783 17.8284 24.7039 18.5153C24.8152 18.6077 24.9072 18.7213 24.9744 18.8494C25.0417 18.9775 25.0829 19.1177 25.0957 19.2618C25.1085 19.406 25.0926 19.5512 25.049 19.6892C25.0053 19.8272 24.9348 19.9551 24.8415 20.0657C24.6524 20.2901 24.3822 20.4306 24.0898 20.4564C23.7975 20.4821 23.5069 20.3912 23.2814 20.2033C22.6154 19.6396 21.9084 19.1262 21.1662 18.6673C20.4446 18.2393 19.6934 17.8634 18.9182 17.5425C17.7309 18.0702 16.4461 18.3428 15.1469 18.3425C13.7146 18.3427 12.3017 18.0114 11.0188 17.3745L10.9516 17.4065C7.83156 18.6161 5.55472 20.4049 4.08429 22.7793C2.60907 25.1617 2.00586 27.8242 2.27626 30.8018C2.30109 31.0933 2.20922 31.3827 2.02082 31.6065C1.83242 31.8303 1.5629 31.9702 1.27144 31.9954C0.980175 32.0221 0.690216 31.9321 0.465225 31.7452C0.240234 31.5583 0.098601 31.2898 0.0714211 30.9986C-0.240585 27.5554 0.469828 24.4177 2.19946 21.6257C3.71949 19.1681 5.95473 17.2577 8.87798 15.8945C7.94657 15.0397 7.20303 14.0007 6.69456 12.8433C6.18609 11.6859 5.9238 10.4354 5.92433 9.17125C5.92433 4.10562 10.0524 0 15.1469 0ZM26.9887 20.3089C27.5967 20.3089 28.0895 20.8049 28.0895 21.4145V24.2177H30.8928C31.0379 24.2173 31.1817 24.2455 31.3159 24.3007C31.4501 24.3559 31.5721 24.437 31.6749 24.5394C31.7778 24.6418 31.8594 24.7634 31.9152 24.8974C31.971 25.0314 31.9998 25.175 32 25.3201C31.9987 25.6127 31.8814 25.8927 31.6738 26.0988C31.4662 26.3049 31.1853 26.4202 30.8928 26.4194H28.0895V29.2242C28.09 29.5167 27.9743 29.7974 27.7679 30.0047C27.5615 30.212 27.2812 30.3289 26.9887 30.3298C26.6962 30.3289 26.4159 30.212 26.2096 30.0047C26.0032 29.7974 25.8875 29.5167 25.8879 29.2242V26.4194H23.0846C22.7921 26.4202 22.5112 26.3049 22.3036 26.0988C22.096 25.8927 21.9787 25.6127 21.9774 25.3201C21.9774 24.7121 22.4734 24.2193 23.0846 24.2193L25.8879 24.2177V21.4145C25.8875 21.122 26.0032 20.8413 26.2096 20.634C26.4159 20.4267 26.6962 20.3098 26.9887 20.3089ZM15.1485 2.20161C11.2764 2.20161 8.13717 5.32163 8.13717 9.17125C8.13717 13.0209 11.2764 16.1425 15.1469 16.1425C19.0174 16.1425 22.1566 13.0225 22.1566 9.17125C22.1566 5.32003 19.019 2.20161 15.1485 2.20161Z" fill="white"/>
-                        </svg>
-                    </a>
-                </li>
-                <li className={`${styles.entry}`} onClick={() => navigate("/admin/staffmembers")}>
-                    <a className={styles.e_icon}>
-                        <svg xmlns="http://www.w3.org/2000/svg"viewBox="0 0 14 18" fill="none">
-                            <path d="M3.5 8.5V7.5H10.5V8.5H3.5ZM3.5 4.5V3.5H10.5V4.5H3.5ZM1 11.5H8.5C8.90667 11.5 9.28 11.5883 9.62 11.765C9.95933 11.941 10.2513 12.1887 10.496 12.508L13 15.758V1.616C13 1.436 12.9423 1.28833 12.827 1.173C12.7117 1.05767 12.5643 1 12.385 1H1.615C1.43567 1 1.28833 1.05767 1.173 1.173C1.05767 1.28833 1 1.436 1 1.616V11.5ZM1.616 17H12.685L9.71 13.117C9.55867 12.9197 9.38 12.7673 9.174 12.66C8.96867 12.5533 8.744 12.5 8.5 12.5H1V16.385C1 16.5643 1.05767 16.7117 1.173 16.827C1.28833 16.9423 1.436 17 1.616 17ZM12.385 18H1.615C1.155 18 0.771 17.846 0.463 17.538C0.155 17.23 0.000666667 16.8457 0 16.385V1.615C0 1.155 0.154333 0.771 0.463 0.463C0.771666 0.155 1.156 0.000666667 1.616 0H12.385C12.845 0 13.2293 0.154333 13.538 0.463C13.8467 0.771667 14.0007 1.156 14 1.616V16.385C14 16.845 13.846 17.2293 13.538 17.538C13.23 17.8467 12.8463 18.0007 12.385 18Z" fill="white"/>
-                        </svg>
-                    </a>
-                </li>
+                {role !== "Manager" && role !== "Designer" ? (
+                    <>
+                        <li className={`${styles.entry}`} onClick={() => navigate(rolePaths.createStaff[role])}>
+                            <a className={styles.e_icon}>
+                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 34 34" fill="none">
+                                    <path d="M15.1469 0C20.2414 0 24.3711 4.10562 24.3711 9.17125C24.3723 10.4987 24.0836 11.8104 23.5254 13.0147C22.9672 14.2191 22.1527 15.2871 21.139 16.1441C21.571 16.3574 21.9646 16.5724 22.3198 16.7889C23.0836 17.2529 23.8783 17.8284 24.7039 18.5153C24.8152 18.6077 24.9072 18.7213 24.9744 18.8494C25.0417 18.9775 25.0829 19.1177 25.0957 19.2618C25.1085 19.406 25.0926 19.5512 25.049 19.6892C25.0053 19.8272 24.9348 19.9551 24.8415 20.0657C24.6524 20.2901 24.3822 20.4306 24.0898 20.4564C23.7975 20.4821 23.5069 20.3912 23.2814 20.2033C22.6154 19.6396 21.9084 19.1262 21.1662 18.6673C20.4446 18.2393 19.6934 17.8634 18.9182 17.5425C17.7309 18.0702 16.4461 18.3428 15.1469 18.3425C13.7146 18.3427 12.3017 18.0114 11.0188 17.3745L10.9516 17.4065C7.83156 18.6161 5.55472 20.4049 4.08429 22.7793C2.60907 25.1617 2.00586 27.8242 2.27626 30.8018C2.30109 31.0933 2.20922 31.3827 2.02082 31.6065C1.83242 31.8303 1.5629 31.9702 1.27144 31.9954C0.980175 32.0221 0.690216 31.9321 0.465225 31.7452C0.240234 31.5583 0.098601 31.2898 0.0714211 30.9986C-0.240585 27.5554 0.469828 24.4177 2.19946 21.6257C3.71949 19.1681 5.95473 17.2577 8.87798 15.8945C7.94657 15.0397 7.20303 14.0007 6.69456 12.8433C6.18609 11.6859 5.9238 10.4354 5.92433 9.17125C5.92433 4.10562 10.0524 0 15.1469 0ZM26.9887 20.3089C27.5967 20.3089 28.0895 20.8049 28.0895 21.4145V24.2177H30.8928C31.0379 24.2173 31.1817 24.2455 31.3159 24.3007C31.4501 24.3559 31.5721 24.437 31.6749 24.5394C31.7778 24.6418 31.8594 24.7634 31.9152 24.8974C31.971 25.0314 31.9998 25.175 32 25.3201C31.9987 25.6127 31.8814 25.8927 31.6738 26.0988C31.4662 26.3049 31.1853 26.4202 30.8928 26.4194H28.0895V29.2242C28.09 29.5167 27.9743 29.7974 27.7679 30.0047C27.5615 30.212 27.2812 30.3289 26.9887 30.3298C26.6962 30.3289 26.4159 30.212 26.2096 30.0047C26.0032 29.7974 25.8875 29.5167 25.8879 29.2242V26.4194H23.0846C22.7921 26.4202 22.5112 26.3049 22.3036 26.0988C22.096 25.8927 21.9787 25.6127 21.9774 25.3201C21.9774 24.7121 22.4734 24.2193 23.0846 24.2193L25.8879 24.2177V21.4145C25.8875 21.122 26.0032 20.8413 26.2096 20.634C26.4159 20.4267 26.6962 20.3098 26.9887 20.3089ZM15.1485 2.20161C11.2764 2.20161 8.13717 5.32163 8.13717 9.17125C8.13717 13.0209 11.2764 16.1425 15.1469 16.1425C19.0174 16.1425 22.1566 13.0225 22.1566 9.17125C22.1566 5.32003 19.019 2.20161 15.1485 2.20161Z" fill="white"/>
+                                </svg>
+                            </a>
+                        </li>
+                    </>
+                    
+                ) : null }
+
+                {role !== "Designer" ?(
+                    <>
+                        <li className={`${styles.entry}`} onClick={() => navigate(rolePaths.viewStaffDetails[role])}>
+                            <a className={styles.e_icon}>
+                                <svg xmlns="http://www.w3.org/2000/svg"viewBox="0 0 14 18" fill="none">
+                                    <path d="M3.5 8.5V7.5H10.5V8.5H3.5ZM3.5 4.5V3.5H10.5V4.5H3.5ZM1 11.5H8.5C8.90667 11.5 9.28 11.5883 9.62 11.765C9.95933 11.941 10.2513 12.1887 10.496 12.508L13 15.758V1.616C13 1.436 12.9423 1.28833 12.827 1.173C12.7117 1.05767 12.5643 1 12.385 1H1.615C1.43567 1 1.28833 1.05767 1.173 1.173C1.05767 1.28833 1 1.436 1 1.616V11.5ZM1.616 17H12.685L9.71 13.117C9.55867 12.9197 9.38 12.7673 9.174 12.66C8.96867 12.5533 8.744 12.5 8.5 12.5H1V16.385C1 16.5643 1.05767 16.7117 1.173 16.827C1.28833 16.9423 1.436 17 1.616 17ZM12.385 18H1.615C1.155 18 0.771 17.846 0.463 17.538C0.155 17.23 0.000666667 16.8457 0 16.385V1.615C0 1.155 0.154333 0.771 0.463 0.463C0.771666 0.155 1.156 0.000666667 1.616 0H12.385C12.845 0 13.2293 0.154333 13.538 0.463C13.8467 0.771667 14.0007 1.156 14 1.616V16.385C14 16.845 13.846 17.2293 13.538 17.538C13.23 17.8467 12.8463 18.0007 12.385 18Z" fill="white"/>
+                                </svg>
+                            </a>
+                        </li>
+                    </>
+                ) : null}
+
+                {role === "Designer" || role === "Worker" ? (
+                    <>
+                        <li className={`${styles.home}`} onClick={() => navigate(rolePaths.attendance[role])}>
+                            <a className={styles.h_icon}>
+                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 34 38" fill="none">
+                                    <path d="M10.7928 5.0787H23.2077C23.557 5.0787 23.8403 4.79548 23.8403 4.44609C23.8403 4.0967 23.557 3.81348 23.2077 3.81348H10.7928C10.4434 3.81348 10.1602 4.0967 10.1602 4.44609C10.1602 4.79548 10.4434 5.0787 10.7928 5.0787Z" fill="white"/>
+                                    <path d="M28.5694 3.81352H27.4248C27.0754 3.81352 26.7922 4.09674 26.7922 4.44614C26.7922 4.79553 27.0754 5.07875 27.4248 5.07875H28.5694C30.8662 5.07875 32.7348 6.94735 32.7348 9.2442V32.3829C32.7348 34.6797 30.8662 36.5483 28.5694 36.5483H5.43064C3.13378 36.5483 1.26518 34.6797 1.26518 32.3829V9.24416C1.26518 6.94731 3.13378 5.0787 5.43064 5.0787H6.57523C6.92463 5.0787 7.20784 4.79548 7.20784 4.44609C7.20784 4.0967 6.92463 3.81348 6.57523 3.81348H5.43064C2.43618 3.81348 0 6.24966 0 9.24411V32.3829C0 35.3773 2.43618 37.8135 5.43064 37.8135H28.5694C31.5638 37.8135 34 35.3773 34 32.3829V9.24416C34 6.2497 31.5638 3.81352 28.5694 3.81352Z" fill="white"/>
+                                    <path d="M8.68339 7.99322C9.03279 7.99322 9.31601 7.71 9.31601 7.36061V1.53105C9.31601 1.18166 9.03279 0.898438 8.68339 0.898438C8.334 0.898438 8.05078 1.18166 8.05078 1.53105V7.36061C8.05078 7.71 8.334 7.99322 8.68339 7.99322Z" fill="white"/>
+                                    <path d="M25.3162 7.99322C25.6656 7.99322 25.9488 7.71 25.9488 7.36061V1.53105C25.9488 1.18166 25.6656 0.898438 25.3162 0.898438C24.9669 0.898438 24.6836 1.18166 24.6836 1.53105V7.36061C24.6835 7.71 24.9668 7.99322 25.3162 7.99322Z" fill="white"/>
+                                    <path d="M13.6935 15.5325C14.4476 15.5325 15.0589 14.9212 15.0589 14.1671C15.0589 13.4131 14.4476 12.8018 13.6935 12.8018C12.9394 12.8018 12.3281 13.4131 12.3281 14.1671C12.3281 14.9212 12.9394 15.5325 13.6935 15.5325Z" fill="white"/>
+                                    <path d="M20.3068 15.5325C21.0609 15.5325 21.6722 14.9212 21.6722 14.1671C21.6722 13.4131 21.0609 12.8018 20.3068 12.8018C19.5527 12.8018 18.9414 13.4131 18.9414 14.1671C18.9414 14.9212 19.5527 15.5325 20.3068 15.5325Z" fill="white"/>
+                                    <path d="M26.9201 15.5325C27.6741 15.5325 28.2854 14.9212 28.2854 14.1671C28.2854 13.4131 27.6741 12.8018 26.9201 12.8018C26.166 12.8018 25.5547 13.4131 25.5547 14.1671C25.5547 14.9212 26.166 15.5325 26.9201 15.5325Z" fill="white"/>
+                                    <path d="M7.08022 20.6185C7.8343 20.6185 8.4456 20.0072 8.4456 19.2531C8.4456 18.499 7.8343 17.8877 7.08022 17.8877C6.32614 17.8877 5.71484 18.499 5.71484 19.2531C5.71484 20.0072 6.32614 20.6185 7.08022 20.6185Z" fill="white"/>
+                                    <path d="M13.6935 20.6185C14.4476 20.6185 15.0589 20.0072 15.0589 19.2531C15.0589 18.499 14.4476 17.8877 13.6935 17.8877C12.9394 17.8877 12.3281 18.499 12.3281 19.2531C12.3281 20.0072 12.9394 20.6185 13.6935 20.6185Z" fill="white"/>
+                                    <path d="M20.3068 20.6185C21.0609 20.6185 21.6722 20.0072 21.6722 19.2531C21.6722 18.499 21.0609 17.8877 20.3068 17.8877C19.5527 17.8877 18.9414 18.499 18.9414 19.2531C18.9414 20.0072 19.5527 20.6185 20.3068 20.6185Z" fill="white"/>
+                                    <path d="M26.9201 20.6185C27.6741 20.6185 28.2854 20.0072 28.2854 19.2531C28.2854 18.499 27.6741 17.8877 26.9201 17.8877C26.166 17.8877 25.5547 18.499 25.5547 19.2531C25.5547 20.0072 26.166 20.6185 26.9201 20.6185Z" fill="white"/>
+                                    <path d="M7.08022 25.7039C7.8343 25.7039 8.4456 25.0926 8.4456 24.3385C8.4456 23.5844 7.8343 22.9731 7.08022 22.9731C6.32614 22.9731 5.71484 23.5844 5.71484 24.3385C5.71484 25.0926 6.32614 25.7039 7.08022 25.7039Z" fill="white"/>
+                                    <path d="M13.6935 25.7039C14.4476 25.7039 15.0589 25.0926 15.0589 24.3385C15.0589 23.5844 14.4476 22.9731 13.6935 22.9731C12.9394 22.9731 12.3281 23.5844 12.3281 24.3385C12.3281 25.0926 12.9394 25.7039 13.6935 25.7039Z" fill="white"/>
+                                    <path d="M20.3068 25.7039C21.0609 25.7039 21.6722 25.0926 21.6722 24.3385C21.6722 23.5844 21.0609 22.9731 20.3068 22.9731C19.5527 22.9731 18.9414 23.5844 18.9414 24.3385C18.9414 25.0926 19.5527 25.7039 20.3068 25.7039Z" fill="white"/>
+                                    <path d="M26.9201 25.7039C27.6741 25.7039 28.2854 25.0926 28.2854 24.3385C28.2854 23.5844 27.6741 22.9731 26.9201 22.9731C26.166 22.9731 25.5547 23.5844 25.5547 24.3385C25.5547 25.0926 26.166 25.7039 26.9201 25.7039Z" fill="white"/>
+                                    <path d="M7.08022 30.7898C7.8343 30.7898 8.4456 30.1785 8.4456 29.4245C8.4456 28.6704 7.8343 28.0591 7.08022 28.0591C6.32614 28.0591 5.71484 28.6704 5.71484 29.4245C5.71484 30.1785 6.32614 30.7898 7.08022 30.7898Z" fill="white"/>
+                                    <path d="M13.6935 30.7898C14.4476 30.7898 15.0589 30.1785 15.0589 29.4245C15.0589 28.6704 14.4476 28.0591 13.6935 28.0591C12.9394 28.0591 12.3281 28.6704 12.3281 29.4245C12.3281 30.1785 12.9394 30.7898 13.6935 30.7898Z" fill="white"/>
+                                    <path d="M20.3068 30.7898C21.0609 30.7898 21.6722 30.1785 21.6722 29.4245C21.6722 28.6704 21.0609 28.0591 20.3068 28.0591C19.5527 28.0591 18.9414 28.6704 18.9414 29.4245C18.9414 30.1785 19.5527 30.7898 20.3068 30.7898Z" fill="white"/>
+                                </svg>
+                            </a>
+                        </li>
+                    </>
+                ) : null}
             </ul>
             <ul className={styles.LogoutContainer}>
                 <li className={styles.Logout}>
