+++ Frontend-Task-page/src/Components/sidebar.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import styles from "../StyleSCSS/sidebar.module.scss";
+import styles from "../StyleSCSS/AdminSideBars.module.scss";
 import MobileSidebar from "./mobileSidebar";
 import { useNavigate } from "react-router-dom";
 
@@ -33,26 +33,27 @@ export default function Sidebar({onSelectSection, onLogout, role}) {
     return(
         <nav className={styles.sidebar}>
             <ul className={styles.Container}>
-                <li className={`${styles.home} ${activeSection === "home" ? styles.active : ""}`} onClick={() => navigate("/user/dashboard")}>
+                <li className={`${styles.entry} ${activeSection === "home" ? styles.active : ""}`} onClick={() => navigate("/user/dashboard")}>
                     <a className={styles.h_icon}>
-                        <svg xmlns="http://www.w3.org/2000/svg" width="32" height="35" viewBox="0 0 32 35" fill="none">
-                            <path d="M0 31V13C0 12.3667 0.142 11.7667 0.426 11.2C0.71 10.6333 1.10133 10.1667 1.6 9.8L13.6 0.8C14.3 0.266667 15.1 0 16 0C16.9 0 17.7 0.266667 18.4 0.8L30.4 9.8C30.9 10.1667 31.292 10.6333 31.576 11.2C31.86 11.7667 32.0013 12.3667 32 13V31C32 32.1 31.608 33.042 30.824 33.826C30.04 34.61 29.0987 35.0013 28 35H22C21.4333 35 20.9587 34.808 20.576 34.424C20.1933 34.04 20.0013 33.5653 20 33V23C20 22.4333 19.808 21.9587 19.424 21.576C19.04 21.1933 18.5653 21.0013 18 21H14C13.4333 21 12.9587 21.192 12.576 21.576C12.1933 21.96 12.0013 22.4347 12 23V33C12 33.5667 11.808 34.042 11.424 34.426C11.04 34.81 10.5653 35.0013 10 35H4C2.9 35 1.95867 34.6087 1.176 33.826C0.393333 33.0433 0.00133333 32.1013 0 31Z" fill="#FF7805"/>
+                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 37 33" fill="none">
+                            <path d="M5.09375 13.1484V31.125C5.09375 31.4482 5.22215 31.7582 5.45071 31.9868C5.67927 32.2153 5.98927 32.3438 6.3125 32.3438H13.625V21.9844C13.625 21.4995 13.8176 21.0345 14.1604 20.6917C14.5033 20.3489 14.9683 20.1562 15.4531 20.1562H21.5469C22.0317 20.1562 22.4967 20.3489 22.8396 20.6917C23.1824 21.0345 23.375 21.4995 23.375 21.9844V32.3438H30.6875C31.0107 32.3438 31.3207 32.2153 31.5493 31.9868C31.7778 31.7582 31.9062 31.4482 31.9062 31.125V13.1484" stroke="white" stroke-linecap="round" stroke-linejoin="round"/>
+                            <path d="M35.5625 16.5L19.3295 0.96095C18.9487 0.558762 18.0582 0.554192 17.6705 0.96095L1.4375 16.5M29.4688 10.6348V1.87501H25.8125V7.13087" stroke="white" stroke-linecap="round" stroke-linejoin="round"/>
                         </svg>
                     </a>
                 </li>
-                <li className={`${styles.attendance} ${activeSection === "attendance" ? styles.active : ""}`} onClick={() => navigate("/user/attendance_logs")}>
-                    <a className={styles.a_icon}>
-                        <svg xmlns="http://www.w3.org/2000/svg" width="40" height="40" viewBox="0 0 40 40" fill="none">
-                            <path opacity="0.5" d="M40 16H0V34C0 35.5913 0.632141 37.1174 1.75736 38.2426C2.88258 39.3679 4.4087 40 6 40H34C35.5913 40 37.1174 39.3679 38.2426 38.2426C39.3679 37.1174 40 35.5913 40 34V16ZM10 12C9.46957 12 8.96086 11.7893 8.58579 11.4142C8.21071 11.0391 8 10.5304 8 10V2C8 1.46957 8.21071 0.960859 8.58579 0.585786C8.96086 0.210714 9.46957 0 10 0C10.5304 0 11.0391 0.210714 11.4142 0.585786C11.7893 0.960859 12 1.46957 12 2V10C12 10.5304 11.7893 11.0391 11.4142 11.4142C11.0391 11.7893 10.5304 12 10 12ZM30 12C29.4696 12 28.9609 11.7893 28.5858 11.4142C28.2107 11.0391 28 10.5304 28 10V2C28 1.46957 28.2107 0.960859 28.5858 0.585786C28.9609 0.210714 29.4696 0 30 0C30.5304 0 31.0391 0.210714 31.4142 0.585786C31.7893 0.960859 32 1.46957 32 2V10C32 10.5304 31.7893 11.0391 31.4142 11.4142C31.0391 11.7893 30.5304 12 30 12Z" fill="#FF7805"/>
-                            <path d="M34 4H32V10C32 10.5304 31.7893 11.0391 31.4142 11.4142C31.0391 11.7893 30.5304 12 30 12C29.4696 12 28.9609 11.7893 28.5858 11.4142C28.2107 11.0391 28 10.5304 28 10V4H12V10C12 10.5304 11.7893 11.0391 11.4142 11.4142C11.0391 11.7893 10.5304 12 10 12C9.46957 12 8.96086 11.7893 8.58579 11.4142C8.21071 11.0391 8 10.5304 8 10V4H6C4.4087 4 2.88258 4.63214 1.75736 5.75736C0.632141 6.88258 0 8.4087 0 10V16H40V10C40 8.4087 39.3679 6.88258 38.2426 5.75736C37.1174 4.63214 35.5913 4 34 4Z" fill="#FF7805"/>
-                        </svg>
-                    </a>
+                <li className={`${styles.entry} ${activeSection === "attendance" ? styles.active : ""}`} onClick={() => navigate("/user/attendance_logs")}>
+                    <a className={styles.e_icon}>
+                                <svg xmlns="http://www.w3.org/2000/svg"viewBox="0 0 14 18" fill="none">
+                                    <path d="M3.5 8.5V7.5H10.5V8.5H3.5ZM3.5 4.5V3.5H10.5V4.5H3.5ZM1 11.5H8.5C8.90667 11.5 9.28 11.5883 9.62 11.765C9.95933 11.941 10.2513 12.1887 10.496 12.508L13 15.758V1.616C13 1.436 12.9423 1.28833 12.827 1.173C12.7117 1.05767 12.5643 1 12.385 1H1.615C1.43567 1 1.28833 1.05767 1.173 1.173C1.05767 1.28833 1 1.436 1 1.616V11.5ZM1.616 17H12.685L9.71 13.117C9.55867 12.9197 9.38 12.7673 9.174 12.66C8.96867 12.5533 8.744 12.5 8.5 12.5H1V16.385C1 16.5643 1.05767 16.7117 1.173 16.827C1.28833 16.9423 1.436 17 1.616 17ZM12.385 18H1.615C1.155 18 0.771 17.846 0.463 17.538C0.155 17.23 0.000666667 16.8457 0 16.385V1.615C0 1.155 0.154333 0.771 0.463 0.463C0.771666 0.155 1.156 0.000666667 1.616 0H12.385C12.845 0 13.2293 0.154333 13.538 0.463C13.8467 0.771667 14.0007 1.156 14 1.616V16.385C14 16.845 13.846 17.2293 13.538 17.538C13.23 17.8467 12.8463 18.0007 12.385 18Z" fill="white"/>
+                                </svg>
+                            </a>
                 </li>
                 {role === "Designer" && (
                     <li className={`${styles.entry} ${activeSection === "entry" ? styles.active : ""}`} onClick={() => navigate("/user/cliententry")}>
                         <a className={styles.e_icon}>
-                            <svg xmlns="http://www.w3.org/2000/svg" width="42" height="42" viewBox="0 0 42 42" fill="none">
-                                <path d="M0.143066 6.57142C0.143066 4.86646 0.82036 3.23133 2.02595 2.02574C3.23154 0.820147 4.86667 0.142853 6.57164 0.142853H31.4288C33.1337 0.142853 34.7689 0.820147 35.9745 2.02574C37.1801 3.23133 37.8574 4.86646 37.8574 6.57142V19.8571C37.0613 19.2587 36.1983 18.7547 35.2859 18.3554V6.57142C35.2859 5.54845 34.8795 4.56737 34.1562 3.84401C33.4328 3.12066 32.4518 2.71428 31.4288 2.71428H6.57164C5.54866 2.71428 4.56758 3.12066 3.84423 3.84401C3.12087 4.56737 2.71449 5.54845 2.71449 6.57142V31.4286C2.71449 32.4515 3.12087 33.4326 3.84423 34.156C4.56758 34.8793 5.54866 35.2857 6.57164 35.2857H18.3556C18.7568 36.2029 19.2625 37.0651 19.8556 37.8571H6.57164C4.86667 37.8571 3.23154 37.1798 2.02595 35.9743C0.82036 34.7687 0.143066 33.1335 0.143066 31.4286V6.57142ZM31.4288 17.2857C31.7698 17.2857 32.0968 17.1503 32.3379 16.9091C32.579 16.668 32.7145 16.341 32.7145 16C32.7145 15.659 32.579 15.332 32.3379 15.0909C32.0968 14.8497 31.7698 14.7143 31.4288 14.7143H18.5716C18.2306 14.7143 17.9036 14.8497 17.6625 15.0909C17.4214 15.332 17.2859 15.659 17.2859 16C17.2859 16.341 17.4214 16.668 17.6625 16.9091C17.9036 17.1503 18.2306 17.2857 18.5716 17.2857H31.4288ZM5.28592 7.42857C5.28592 7.08757 5.42138 6.76055 5.6625 6.51943C5.90362 6.27831 6.23065 6.14285 6.57164 6.14285H31.4288C31.7698 6.14285 32.0968 6.27831 32.3379 6.51943C32.579 6.76055 32.7145 7.08757 32.7145 7.42857C32.7145 7.76956 32.579 8.09659 32.3379 8.3377C32.0968 8.57882 31.7698 8.71428 31.4288 8.71428H6.57164C6.23065 8.71428 5.90362 8.57882 5.6625 8.3377C5.42138 8.09659 5.28592 7.76956 5.28592 7.42857ZM9.57164 12.1429C8.435 12.1429 7.34491 12.5944 6.54118 13.3981C5.73745 14.2018 5.28592 15.2919 5.28592 16.4286C5.28592 17.5652 5.73745 18.6553 6.54118 19.459C7.34491 20.2628 8.435 20.7143 9.57164 20.7143C10.7083 20.7143 11.7984 20.2628 12.6021 19.459C13.4058 18.6553 13.8574 17.5652 13.8574 16.4286C13.8574 15.2919 13.4058 14.2018 12.6021 13.3981C11.7984 12.5944 10.7083 12.1429 9.57164 12.1429ZM7.85735 16.4286C7.85735 15.9739 8.03796 15.5379 8.35946 15.2164C8.68095 14.8949 9.11698 14.7143 9.57164 14.7143C10.0263 14.7143 10.4623 14.8949 10.7838 15.2164C11.1053 15.5379 11.2859 15.9739 11.2859 16.4286C11.2859 16.8832 11.1053 17.3193 10.7838 17.6408C10.4623 17.9622 10.0263 18.1429 9.57164 18.1429C9.11698 18.1429 8.68095 17.9622 8.35946 17.6408C8.03796 17.3193 7.85735 16.8832 7.85735 16.4286ZM5.28592 28.4286C5.28592 27.2919 5.73745 26.2018 6.54118 25.3981C7.34491 24.5944 8.435 24.1429 9.57164 24.1429C10.7083 24.1429 11.7984 24.5944 12.6021 25.3981C13.4058 26.2018 13.8574 27.2919 13.8574 28.4286C13.8574 29.5652 13.4058 30.6553 12.6021 31.459C11.7984 32.2628 10.7083 32.7143 9.57164 32.7143C8.435 32.7143 7.34491 32.2628 6.54118 31.459C5.73745 30.6553 5.28592 29.5652 5.28592 28.4286ZM9.57164 26.7143C9.11698 26.7143 8.68095 26.8949 8.35946 27.2164C8.03796 27.5379 7.85735 27.9739 7.85735 28.4286C7.85735 28.8832 8.03796 29.3193 8.35946 29.6408C8.68095 29.9622 9.11698 30.1429 9.57164 30.1429C10.0263 30.1429 10.4623 29.9622 10.7838 29.6408C11.1053 29.3193 11.2859 28.8832 11.2859 28.4286C11.2859 27.9739 11.1053 27.5379 10.7838 27.2164C10.4623 26.8949 10.0263 26.7143 9.57164 26.7143ZM41.2859 30.1429C41.2859 33.0981 40.1119 35.9324 38.0223 38.022C35.9326 40.1117 33.0983 41.2857 30.1431 41.2857C27.1878 41.2857 24.3536 40.1117 22.2639 38.022C20.1742 35.9324 19.0002 33.0981 19.0002 30.1429C19.0002 27.1876 20.1742 24.3534 22.2639 22.2637C24.3536 20.174 27.1878 19 30.1431 19C33.0983 19 35.9326 20.174 38.0223 22.2637C40.1119 24.3534 41.2859 27.1876 41.2859 30.1429ZM31.0002 23.2857C31.0002 23.0584 30.9099 22.8404 30.7492 22.6796C30.5884 22.5189 30.3704 22.4286 30.1431 22.4286C29.9157 22.4286 29.6977 22.5189 29.537 22.6796C29.3762 22.8404 29.2859 23.0584 29.2859 23.2857V29.2857H23.2859C23.0586 29.2857 22.8406 29.376 22.6798 29.5368C22.5191 29.6975 22.4288 29.9155 22.4288 30.1429C22.4288 30.3702 22.5191 30.5882 22.6798 30.7489C22.8406 30.9097 23.0586 31 23.2859 31H29.2859V37C29.2859 37.2273 29.3762 37.4453 29.537 37.6061C29.6977 37.7668 29.9157 37.8571 30.1431 37.8571C30.3704 37.8571 30.5884 37.7668 30.7492 37.6061C30.9099 37.4453 31.0002 37.2273 31.0002 37V31H37.0002C37.2275 31 37.4456 30.9097 37.6063 30.7489C37.767 30.5882 37.8574 30.3702 37.8574 30.1429C37.8574 29.9155 37.767 29.6975 37.6063 29.5368C37.4456 29.376 37.2275 29.2857 37.0002 29.2857H31.0002V23.2857Z" fill="#FF7805"/>
+                            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32" fill="none">
+                                <path d="M22.4004 0.5C22.4798 0.500103 22.5561 0.531722 22.6123 0.587891C22.6685 0.644096 22.7001 0.720355 22.7002 0.799805V3.7002H28C28.9283 3.7002 29.8182 4.06921 30.4746 4.72559C31.1309 5.38195 31.5 6.27198 31.5 7.2002V28C31.5 28.9283 31.131 29.8182 30.4746 30.4746C29.8182 31.131 28.9283 31.5 28 31.5H4.00098C3.0729 31.4989 2.18262 31.1299 1.52637 30.4736C0.911122 29.8584 0.548749 29.0377 0.504883 28.1729L0.5 27.999V7.2002L0.504883 7.02734C0.548702 6.16244 0.911136 5.34184 1.52637 4.72656C2.14162 4.11131 2.96225 3.74796 3.82715 3.7041L4.00098 3.7002H9.2998V0.799805C9.29986 0.720357 9.33154 0.644096 9.3877 0.587891C9.44386 0.531722 9.52019 0.500103 9.59961 0.5C9.67917 0.5 9.75624 0.53163 9.8125 0.587891C9.86859 0.644086 9.90034 0.720401 9.90039 0.799805V3.7002H22.0996V0.799805C22.0997 0.720401 22.1314 0.644086 22.1875 0.587891C22.2438 0.53163 22.3208 0.5 22.4004 0.5ZM1.09961 28C1.09961 28.7691 1.40536 29.5069 1.94922 30.0508C2.49307 30.5946 3.23087 30.9004 4 30.9004H28C28.7691 30.9004 29.5069 30.5946 30.0508 30.0508C30.5946 29.5069 30.9004 28.7691 30.9004 28V13.9004H1.09961V28ZM4 4.2998C2.39542 4.2998 1.09961 5.59969 1.09961 7.2002V13.2998H30.9004V7.2002C30.9004 6.43111 30.5946 5.69326 30.0508 5.14941C29.5069 4.60556 28.7691 4.2998 28 4.2998H22.7002V7.2002C22.7001 7.27964 22.6685 7.3559 22.6123 7.41211C22.5561 7.46828 22.4798 7.4999 22.4004 7.5C22.3208 7.5 22.2438 7.46837 22.1875 7.41211C22.1314 7.35591 22.0997 7.2796 22.0996 7.2002V4.2998H9.90039V7.2002C9.90034 7.2796 9.86859 7.35591 9.8125 7.41211C9.75624 7.46837 9.67917 7.5 9.59961 7.5C9.52019 7.4999 9.44386 7.46828 9.3877 7.41211C9.33154 7.3559 9.29986 7.27964 9.2998 7.2002V4.2998H4Z" fill="white" stroke="white"/>
+                                <path d="M15.3924 17.2001H15.4084V17.2171H15.3924V17.2001ZM15.3924 21.9916H15.4084V22.0086H15.3924V21.9916ZM15.3924 26.783H15.4084V26.8001H15.3924V26.783ZM10.8964 21.9916H10.9124V22.0086H10.8964V21.9916ZM10.8964 26.783H10.9124V26.8001H10.8964V26.783ZM6.40039 21.9916H6.41638V22.0086H6.40039V21.9916ZM6.40039 26.783H6.41638V26.8001H6.40039V26.783ZM19.8884 17.2001H19.9044V17.2171H19.8884V17.2001ZM19.8884 21.9916H19.9044V22.0086H19.8884V21.9916ZM19.8884 26.783H19.9044V26.8001H19.8884V26.783ZM24.3844 17.2001H24.4004V17.2171H24.3844V17.2001ZM24.3844 21.9916H24.4004V22.0086H24.3844V21.9916Z" fill="white" stroke="white" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/>
                             </svg>
                         </a>
                     </li>
@@ -61,9 +62,9 @@ export default function Sidebar({onSelectSection, onLogout, role}) {
             <ul className={styles.LogoutContainer}>
                 <li className={styles.Logout}>
                     <button className={styles.l_icon} onClick={onLogout}>
-                        <svg xmlns="http://www.w3.org/2000/svg" width="49" height="49" viewBox="0 0 49 49" fill="none">
-                            <path d="M13.2708 7.65625C12.1969 7.65625 10.7188 8.9425 10.7188 11.3741V37.6279C10.7188 40.0616 12.1969 41.3458 13.2708 41.3458H27.5543C27.9604 41.3458 28.3499 41.5071 28.6371 41.7943C28.9243 42.0815 29.0856 42.4709 29.0856 42.877C29.0856 43.2832 28.9243 43.6726 28.6371 43.9598C28.3499 44.247 27.9604 44.4083 27.5543 44.4083H13.2708C9.83471 44.4083 7.65625 40.9926 7.65625 37.6279V11.3721C7.65625 8.00742 9.83471 4.59171 13.2708 4.59171H27.5625C27.9686 4.59171 28.3581 4.75304 28.6453 5.0402C28.9324 5.32737 29.0938 5.71685 29.0938 6.12296C29.0938 6.52907 28.9324 6.91855 28.6453 7.20572C28.3581 7.49288 27.9686 7.65421 27.5625 7.65421L13.2708 7.65625Z" fill="#de110eff"/>
-                            <path d="M33.7489 16.2721C33.4618 15.9853 33.0726 15.8243 32.6668 15.8243C32.261 15.8243 31.8719 15.9853 31.5847 16.2721V22.9606H19.396C18.9899 22.9606 18.6004 23.1219 18.3132 23.4091C18.0261 23.6962 17.8647 24.0857 17.8647 24.4918C17.8647 24.8979 18.0261 25.2874 18.3132 25.5746C18.6004 25.8618 18.9899 26.0231 19.396 26.0231H31.5847V32.7279C31.8719 33.0147 32.261 33.1757 32.6668 33.1757C33.0726 33.1757 33.4618 33.0147 33.7489 32.7279L40.8947 25.5821C41.0375 25.4395 41.1506 25.2702 41.2277 25.0838C41.3048 24.8974 41.3443 24.6976 41.3439 24.4959V24.4918C41.3438 24.2679 41.2945 24.0467 41.1996 23.8439C41.1047 23.641 40.9665 23.4615 40.7947 23.3179L33.7489 16.2721Z" fill="#de110eff"/>
+                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 49 49" fill="none">
+                            <path d="M13.2708 7.65625C12.1969 7.65625 10.7188 8.9425 10.7188 11.3741V37.6279C10.7188 40.0616 12.1969 41.3458 13.2708 41.3458H27.5543C27.9604 41.3458 28.3499 41.5071 28.6371 41.7943C28.9243 42.0815 29.0856 42.4709 29.0856 42.877C29.0856 43.2832 28.9243 43.6726 28.6371 43.9598C28.3499 44.247 27.9604 44.4083 27.5543 44.4083H13.2708C9.83471 44.4083 7.65625 40.9926 7.65625 37.6279V11.3721C7.65625 8.00742 9.83471 4.59171 13.2708 4.59171H27.5625C27.9686 4.59171 28.3581 4.75304 28.6453 5.0402C28.9324 5.32737 29.0938 5.71685 29.0938 6.12296C29.0938 6.52907 28.9324 6.91855 28.6453 7.20572C28.3581 7.49288 27.9686 7.65421 27.5625 7.65421L13.2708 7.65625Z" fill="#d4652eff"/>
+                            <path d="M33.7489 16.2721C33.4618 15.9853 33.0726 15.8243 32.6668 15.8243C32.261 15.8243 31.8719 15.9853 31.5847 16.2721V22.9606H19.396C18.9899 22.9606 18.6004 23.1219 18.3132 23.4091C18.0261 23.6962 17.8647 24.0857 17.8647 24.4918C17.8647 24.8979 18.0261 25.2874 18.3132 25.5746C18.6004 25.8618 18.9899 26.0231 19.396 26.0231H31.5847V32.7279C31.8719 33.0147 32.261 33.1757 32.6668 33.1757C33.0726 33.1757 33.4618 33.0147 33.7489 32.7279L40.8947 25.5821C41.0375 25.4395 41.1506 25.2702 41.2277 25.0838C41.3048 24.8974 41.3443 24.6976 41.3439 24.4959V24.4918C41.3438 24.2679 41.2945 24.0467 41.1996 23.8439C41.1047 23.641 40.9665 23.4615 40.7947 23.3179L33.7489 16.2721Z" fill="#d4652eff"/>
                         </svg>
                     </button>
                 </li>
