+++ Frontend-Task-page/src/Pages/Manager_Route.jsx
@@ -1,26 +1,24 @@
 import React from 'react';
 import { Routes, Route, Navigate, useNavigate } from "react-router-dom";
-//import { useEffect, useState } from 'react';
-import AdminHeader from "../Components/header.jsx";
+import { useEffect, useState } from 'react';
+import HRHeader from "../Components/header.jsx";
 import StaffUsers from '../Components/userList.jsx';
-import AdminSideBar from "../Components/AdminSidebar.jsx";
-import AdminClientView from "../Components/ClientEntry.jsx"
-import SignupStaff from "../Components/Signup.jsx";
+import HRSideBar from "../Components/AdminSidebar.jsx";
+import ManagerClientView from "../Components/ClientEntry.jsx"
+
 import StaffAssigns from "../Components/Staff_Admin.jsx"
 import styles from "../StyleSCSS/admin.module.scss"
 
-
-export default function Admin({ user, onLogout }) {
+export default function managerStaff({ user, onLogout }) {
     return(
         <div className={styles.admin_layout}>
-            <AdminHeader/>
+            <HRHeader/>
             <div className={styles.admin_main}>
-                <AdminSideBar onLogout={onLogout}/>
+                <HRSideBar onLogout={onLogout}/>
                 <div className={styles.admin_content}>
                     <Routes>
                         <Route path="/Home" element={<StaffAssigns />} />
-                        <Route path="/clientView" element={<AdminClientView />} />
-                        <Route path='/createStaff' element={<SignupStaff/>}/>
+                        <Route path="/clientView" element={<ManagerClientView />} />
                         <Route path='/viewStaffDetails' element={<StaffUsers/>}/>
                         <Route path="*" element={<Navigate to="Home" replace />} />
                     </Routes>
