+++ Frontend-Task-page/src/Pages/UserRoute.jsx
@@ -3,7 +3,7 @@ import { Routes, Route, Navigate, useNavigate } from "react-router-dom";
 import { useEffect, useState } from 'react';
 import styles from "../StyleSCSS/user.module.scss"
 import UserHeader from "../Components/header.jsx";
-import UserSideBar from "../Components/sidebar.jsx";
+import UserSideBar from "../Components/AdminSidebar.jsx";
 import DashBoard from "../Components/HomeComponent.jsx"
 import AttendanceLog from "../Components/Attendance.jsx"
 import ClientEntry from "../Components/ClientEntry.jsx"
@@ -17,10 +17,10 @@ export default function UserPage({ onLogout }) {
                 <UserSideBar role={role} onLogout={onLogout}/>
                 <div className={styles.user_content}>
                     <Routes>
-                        <Route path='/dashboard' element={<DashBoard role={role}/>}/>
-                        <Route path='/attendance_logs' element={<AttendanceLog/>}/>
-                        <Route path='/cliententry' element={<ClientEntry/>}/>
-                        <Route path='*' element={<Navigate to="dashboard" replace />}/>
+                        <Route path='/Home' element={<DashBoard role={role}/>}/>
+                        <Route path='/attendance' element={<AttendanceLog/>}/>
+                        <Route path='/clientView' element={<ClientEntry/>}/>
+                        <Route path='*' element={<Navigate to="Home" replace />}/>
                     </Routes>
                 </div>
             </div>
