+++ Frontend-Task-page/src/StyleSCSS/CalendarModal.module.scss
@@ -65,10 +65,11 @@
   background-color: #fff;
   color: #222;
   border-radius: 8px;
-  box-shadow: 0 6px 12px rgba(0, 0, 0, 0.1);
+  //box-shadow: 0 6px 12px rgba(0, 0, 0, 0.1);
   font-family: Arial, sans-serif;
   line-height: 1.125em;
   padding: 1rem;
+  border: none !important;
 }
 
 :global(.react-calendar__tile--active) {
@@ -83,4 +84,10 @@
   border-radius: 6px;
   color: #6f48eb;
   font-weight: bold;
+}
+
+@media (max-width : 500px) {
+  .calendar_modal{
+    padding: 1rem;
+  }
 }
