+++ Frontend-Task-page/src/StyleSCSS/Client.module.scss
@@ -248,7 +248,7 @@
 
         .client {
             font-size: 0.8rem;
-            padding: 10px;
+            padding: 5px;
         }
     }
 }
