+++ Frontend-Task-page/src/StyleSCSS/admin.module.scss
@@ -60,6 +60,7 @@
   width: 100%;
   overflow-x: auto;
   margin: 0 auto;
+  
 
   .loader_wrapper{
     position: fixed; /* take up full screen */
@@ -161,6 +162,7 @@
         }
       }
     }
+    
     .staffTable__userConatiiner{
       display: flex;
       width: 100%;
