+++ Frontend-Task-page/src/StyleSCSS/dashboard.module.scss
@@ -5,7 +5,7 @@
     //padding: 20px;
     justify-content: space-between;
     align-items: center;
-    gap: clamp(10px, 10px, 25px);
+    gap: clamp(10px, 2vw, 25px);
 
     .dashboardSub{
         display: flex;
@@ -13,7 +13,7 @@
         align-items: center;
         justify-content: space-between;
         width: 100%;
-        background-color: #FF9B46;
+        background-color: #121212;
         box-shadow: 0px 4px 7.4px 0px rgba(0, 0, 0, 0.1);
         //min-width: 350px !important;
         border-radius: 10px;
@@ -28,18 +28,18 @@
     min-height: 100px !important;
     justify-content: center;
     align-items: flex-start;
-    padding: clamp(5px, 12px, 20px);
+    padding: clamp(5px, 2vw, 20px);
     
     
 
     .text{
         width: 100%;
         height: 100%;
-        color: #333333;
+        color: #F2F0EF;
         margin-right: 20px;
 
         h1{
-            font-size: 5rem;
+            font-size: 4rem;
             font-weight: 500;
         }
