+++ Frontend-Task-page/src/StyleSCSS/header.module.scss
@@ -37,7 +37,7 @@
   justify-content: space-between;
   background-color: #1F1F1F;
   border-radius: 100px;
-  padding: 0 0 0 16px ;
+  padding: 0 0 0 25px ;
   height: 60px;
   min-width: max-content;
   width: fit-content;
@@ -51,14 +51,14 @@
     color: white;
 
     .EmployeeId {
-      font-size: clamp(14px, 2vw, 1rem);
+      font-size: clamp(10px, 2vw, 1rem);
       font-weight: 600;
       line-height: 1.2;
       color: #ffffff;
     }
 
     .roles {
-      font-size: clamp(0.7rem, 1.8vw, 1rem);
+      font-size: clamp(0.6rem, 1.8vw, 1rem);
       font-weight: 400;
       opacity: 0.8;
     }
@@ -89,11 +89,8 @@
 
   .ProfileContainer{
     height: auto;
-    .nametagContainer{
-      .EmployeeId{
-        
-      }
-    }
+    padding: 0 0 0 15px;
+    gap: 8px;
 
     .EmployeeIcon {
       width: 40px;
