+++ Frontend-Task-page/src/StyleSCSS/user.module.scss
@@ -22,7 +22,7 @@
             min-width: 0;
             display: flex;
             flex-direction: column;
-            background-color: #FF7805;
+            //background-color: #FF7805;
             overflow-y: auto;
             min-height: 0;
             max-height: 100vh;
