+++ Frontend-Task-page/src/main.jsx
@@ -6,9 +6,9 @@ import AppRoutes from "./App.jsx"; // renamed from App to AppRoutes if needed
 import "./index.css"; // your styling
 
 ReactDOM.createRoot(document.getElementById("root")).render(
-  <React.StrictMode>
+  //<React.StrictMode>
     <BrowserRouter>
       <AppRoutes />
     </BrowserRouter>
-  </React.StrictMode>
+  //</React.StrictMode>
 );
